Improve performance of worldometer api
When a metric was requested by worldometer.api, a double rendering
occurred. The first when the object is instantiated and the second when
a metric is requested. Then, only after these two surrenders,
did I obtain the requested metric. This increased the execution
time +10~20 seconds for each requested metric.

Now, only rendering when the object is instantiated occurs.

The user will have to explicitly call the update_metrics function to
update the data stored in the instance object of the Worldometer class.

This brings a very significant performance gain, practically zeroing
the execution time for each requested metric.